Meteors now NAPI compatible!

+++ com/github/StormTeam/Storm/GlobalVariables.java
@@ -84,12 +84,12 @@ public class GlobalVariables extends ReflectConfiguration {
     };
     //Natural Disasters
     //-Meteors
-    public double Natural__Disasters_Meteor_Chance__To__Spawn = 8;
+    public int Natural__Disasters_Meteor_Chance__To__Spawn = 8;
     public String Natural__Disasters_Meteor_Message__On__Meteor__Crash = "A meteor has exploded at %x, %y, %z.";
     public int Natural__Disasters_Meteor_Shockwave_Damage = 10;
     public int Natural__Disasters_Meteor_Shockwave_Damage__Radius = 100;
     public String Natural__Disasters_Meteor_Shockwave_Damage__Message = "You have been flattened by a meteor!";
-    public long Natural__Disasters_Meteor_Scheduler__Recalculation__Intervals__In__Ticks = 72000;
+    public int Natural__Disasters_Meteor_Scheduler__Recalculation__Intervals__In__Ticks = 72000;
     public boolean Natural__Disasters_Meteor_Meteor_Spawn = true;
     //-Wildfires
     public int Natural__Disasters_Wildfires_Chance__To__Start = 20;
